+++ .gitignore
@@ -1,0 +1,3 @@
+*.pyc
+.git/
+__pycache__/
