+++ data/input/fault_model_2.png
