+++ data/input/marmousi_model.jpg
