+++ data/input/paint_test.png
