+++ src/routines/predominant_brightness.py
@@ -1,0 +1,33 @@
+class ModelRoutine(CreateModel):
+    """
+    Extra routine for images not made in MS Paint
+    """
+    def __init__(self, model):
+        self.model = model
+        self.height, self.width = self.model.shape
+
+    def model_routine_loop(self):
+        for i in range(self.height):
+            for j in range(self.width):
+                adj_values = self.__get_adjacent(self.model, i, j)
+                arr_condition = self.__check_diff_brightness_condition(adj_values)
+                if len(arr_condition) > 0:
+                    self.model[i][j] = arr_condition[0]
+        return self.model
+
+    def __get_adjacent(self, arr: list, i: int, j: int) -> list:
+        # I may or may not had a little help on this one :running:
+        rows = len(arr)
+        cols = len(arr[0])
+
+        v = []
+        for k in range(max(0, i-1), min(i+2, rows)):
+            for l in range(max(0, j-1), min(j+2, cols)):
+                if (k != i or l != j) and arr[k][l] != arr[i][j]:
+                    v.append(arr[k][l])
+        return v
+
+    def __check_diff_brightness_condition(self, adj_arr: list):
+        unique_values, counts = np.unique(adj_arr, return_counts=True)
+        arr_brightness_condition = unique_values[counts > 5]
+        return arr_brightness_condition if len(arr_brightness_condition) > 0 else []
